RF - renamed scil_tractogram_filter.py
