+++ scripts/legacy/scil_filter_tractogram.py
@@ -2,11 +2,11 @@
 # -*- coding: utf-8 -*-
 
 from scilpy.io.deprecator import deprecate_script
-from scripts.scil_tractogram_filter import main as new_main
+from scripts.scil_tractogram_filter_by_roi import main as new_main
 
 
 DEPRECATION_MSG = """
-This script has been renamed scil_tractogram_filter.py. Please change
+This script has been renamed scil_tractogram_filter_by_roi.py. Please change
 your existing pipelines accordingly.
 """
