+++ scripts/scil_tractogram_filter_by_roi.py
